% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-space.R
\name{add_model}
\alias{add_model}
\title{Add Model to Space}
\usage{
add_model(model_space, model, prior_weight = NULL)
}
\arguments{
\item{model_space}{A \code{bs_model_space} object}

\item{model}{A \code{bs_model} object to add}

\item{prior_weight}{Prior probability for the new model. The existing
priors are rescaled to accommodate.}
}
\value{
Updated \code{bs_model_space}
}
\description{
Adds a new model to an existing model space.
}
\examples{
space <- model_space(bs_model_uniform())
space <- add_model(space, bs_model_gaussian(), prior_weight = 0.3)
}
