% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-gaussian.R
\name{bs_model_gaussian_mixture}
\alias{bs_model_gaussian_mixture}
\title{Create Multi-Modal Gaussian Mixture Model}
\usage{
bs_model_gaussian_mixture(means, sds, weights = NULL, name = NULL)
}
\arguments{
\item{means}{Vector of means for each component}

\item{sds}{Vector of standard deviations for each component}

\item{weights}{Vector of mixture weights (must sum to 1)}

\item{name}{Optional name for the model}
}
\value{
A \code{bs_model_gaussian_mixture} object
}
\description{
Creates a model based on a mixture of Gaussians for multi-modal data.
}
\examples{
# Two-component mixture
model <- bs_model_gaussian_mixture(
  means = c(50, 150),
  sds = c(10, 20),
  weights = c(0.7, 0.3)
)
}
