% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-geom.R
\name{geom_surprise_density}
\alias{geom_surprise_density}
\title{Surprise Density Plot}
\usage{
geom_surprise_density(
  mapping = NULL,
  data = NULL,
  which = c("surprise", "signed_surprise"),
  fill = "#4575b4",
  color = "black",
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{data}{Data with surprise values}

\item{which}{Which surprise to plot: "surprise" or "signed_surprise"}

\item{fill}{Fill color}

\item{color}{Border color}

\item{...}{Additional arguments passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}}
}
\value{
A ggplot2 layer
}
\description{
Creates a density plot of surprise values.
}
