% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{get_model_space}
\alias{get_model_space}
\title{Get the model space from a surprise result}
\usage{
get_model_space(x, ...)
}
\arguments{
\item{x}{A \code{bs_surprise} or \code{bs_surprise_sf} object}

\item{...}{Additional arguments (unused)}
}
\value{
A \code{bs_model_space} object
}
\description{
Get the model space from a surprise result
}
