% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{get_surprise}
\alias{get_surprise}
\title{Extract surprise values from result objects}
\usage{
get_surprise(x, type = c("surprise", "signed"), ...)
}
\arguments{
\item{x}{A \code{bs_surprise}, \code{bs_surprise_sf}, or \code{bs_surprise_temporal} object}

\item{type}{Which surprise to extract: "surprise" or "signed"}

\item{...}{Additional arguments (unused)}
}
\value{
Numeric vector of surprise values
}
\description{
Extract surprise values from result objects
}
