% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-space.R
\name{model_names}
\alias{model_names}
\title{Get Model Names}
\usage{
model_names(model_space)
}
\arguments{
\item{model_space}{A \code{bs_model_space} object}
}
\value{
Character vector of model names
}
\description{
Get Model Names
}
