% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-space.R
\name{n_models}
\alias{n_models}
\title{Get Number of Models}
\usage{
n_models(model_space)
}
\arguments{
\item{model_space}{A \code{bs_model_space} object}
}
\value{
Integer number of models
}
\description{
Get Number of Models
}
