% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-space.R
\name{remove_model}
\alias{remove_model}
\title{Remove Model from Space}
\usage{
remove_model(model_space, which)
}
\arguments{
\item{model_space}{A \code{bs_model_space} object}

\item{which}{Integer index or character name of model to remove}
}
\value{
Updated \code{bs_model_space}
}
\description{
Removes a model from an existing model space.
}
