% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scale.R
\name{scale_fill_surprise_diverging_binned}
\alias{scale_fill_surprise_diverging_binned}
\title{Binned Diverging Surprise Scale}
\usage{
scale_fill_surprise_diverging_binned(
  n.breaks = 7,
  palette = "RdBu",
  direction = -1,
  name = "Signed Surprise",
  ...
)
}
\arguments{
\item{n.breaks}{Number of breaks/bins}

\item{palette}{ColorBrewer diverging palette name.
Options: "RdBu", "RdYlBu", "BrBG", "PiYG", "PRGn", "PuOr", "RdGy", "Spectral"}

\item{direction}{Direction of palette (1 = normal, -1 = reversed)}

\item{name}{Legend title}

\item{...}{Additional arguments}
}
\value{
A ggplot2 scale object
}
\description{
Binned diverging scale for signed surprise values.
}
