% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-space.R
\name{set_prior}
\alias{set_prior}
\title{Set Prior Probabilities}
\usage{
set_prior(model_space, prior)
}
\arguments{
\item{model_space}{A \code{bs_model_space} object}

\item{prior}{Numeric vector of new prior probabilities (must sum to 1)}
}
\value{
Updated \code{bs_model_space}
}
\description{
Updates the prior probabilities of a model space.
}
