% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-methods.R
\name{st_density_at}
\alias{st_density_at}
\title{Evaluate Density at sf Feature Locations}
\usage{
st_density_at(density, x)
}
\arguments{
\item{density}{Density estimate from \code{\link[=st_density]{st_density()}}}

\item{x}{sf object with features to evaluate at}
}
\value{
Numeric vector of density values
}
\description{
Evaluates a density estimate at the locations of features in an sf object.
}
