% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-stat.R
\docType{data}
\name{StatSurprise}
\alias{StatSurprise}
\title{StatSurprise ggproto Object}
\format{
A ggproto object.
}
\usage{
StatSurprise
}
\description{
StatSurprise ggproto Object
}
\keyword{internal}
