% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-stat.R
\docType{data}
\name{StatSurpriseSf}
\alias{StatSurpriseSf}
\title{StatSurpriseSf ggproto Object (for sf integration)}
\format{
A ggproto object.
}
\usage{
StatSurpriseSf
}
\description{
A specialized stat that works with sf geometries.
}
\keyword{internal}
